+++ src/main/java/com/example/MtaAPI/service/serviceImpl/ProductServiceImpl.java
@@ -6,6 +6,7 @@ import com.example.MtaAPI.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.annotation.Resource;
 import java.util.List;
 import java.util.Optional;
 
@@ -20,7 +21,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public Optional<Product> findById(Integer id) {
+    public Optional<Product> findById(Long id) {
         return productRepository.findById(id);
     }
